refactor(data): extract DataExportImport shared models and helpers
Move DataExportImport domain types, defaults, and status/format helpers into dedicated modules so the component stays focused on UI flow and orchestration.
