+++ src/components/data-export-import/utils.tsx
@@ -1,0 +1,53 @@
+import {
+  CheckCircle as CheckCircleIcon,
+  Error as ErrorIcon,
+  Info as InfoIcon,
+} from "@mui/icons-material";
+import { LinearProgress } from "@mui/material";
+
+type MuiColor =
+  | "default"
+  | "primary"
+  | "secondary"
+  | "error"
+  | "info"
+  | "success"
+  | "warning";
+
+export function getStatusIcon(status: string) {
+  switch (status) {
+    case "completed":
+      return <CheckCircleIcon color="success" />;
+    case "failed":
+      return <ErrorIcon color="error" />;
+    case "in_progress":
+      return <LinearProgress />;
+    default:
+      return <InfoIcon color="inherit" />;
+  }
+}
+
+export function getStatusColor(status: string): MuiColor {
+  switch (status) {
+    case "completed":
+      return "success";
+    case "failed":
+      return "error";
+    case "in_progress":
+      return "info";
+    default:
+      return "default";
+  }
+}
+
+export function formatFileSize(bytes: number): string {
+  if (bytes === 0) return "0 Bytes";
+  const k = 1024;
+  const sizes = ["Bytes", "KB", "MB", "GB"];
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+  return `${parseFloat((bytes / Math.pow(k, i)).toFixed(2))} ${sizes[i]}`;
+}
+
+export function formatDate(dateString: string): string {
+  return new Date(dateString).toLocaleString("fr-FR");
+}
